Struct ethers_contract::Abigen

pub struct Abigen { /* private fields */ }

Programmatically generate type-safe Rust bindings for an Ethereum smart contract from its ABI.

For all the supported ABI sources, see Source.

To generate bindings for multiple contracts at once, see MultiAbigen.

To generate bindings at compile time, see the abigen! macro, or use in a build.rs file.

Example

Running the code below will generate a file called token.rs containing the bindings inside, which exports an ERC20Token struct, along with all its events.

use ethers_contract_abigen::Abigen;

Abigen::new("ERC20Token", "./abi.json")?.generate()?.write_to_file("token.rs")?;

Implementations

impl Abigen

pub fn new<T, S>(contract_name: T, abi_source: S) -> Result<Abigen, Report>where T: AsRef<str>, S: AsRef<str>,

Creates a new builder with the given contract name and ABI source strings.

Errors

If contract_name could not be parsed as a valid Ident, or if abi_source could not be parsed as a valid Source.

pub fn new_raw(contract_name: Ident, abi_source: Source) -> Abigen

Creates a new builder with the given contract name Ident and ABI source.

pub fn from_file(path: impl AsRef<Path>) -> Result<Abigen, Report>

Attempts to load a new builder from an ABI JSON file at the specific path.

pub fn add_event_alias<S1, S2>(self, signature: S1, alias: S2) -> Abigenwhere S1: Into<String>, S2: Into<String>,

Manually adds a solidity event alias to specify what the event struct and function name will be in Rust.

For events without an alias, the PascalCase event name will be used.

pub fn add_method_alias<S1, S2>(self, signature: S1, alias: S2) -> Abigenwhere S1: Into<String>, S2: Into<String>,

Add a Solidity method error alias to specify the generated method name.

For methods without an alias, the snake_case method name will be used.

pub fn add_error_alias<S1, S2>(self, signature: S1, alias: S2) -> Abigenwhere S1: Into<String>, S2: Into<String>,

Add a Solidity custom error alias to specify the generated struct’s name.

For errors without an alias, the PascalCase error name will be used.

pub fn add_derive<S>(self, derive: S) -> Result<Abigen, Report>where S: AsRef<str>,

Add a custom derive to the derives for all structs and enums.

For example, this makes it possible to derive serde::Serialize and serde::Deserialize.

pub fn format(self, format: bool) -> Abigen

Specify whether to format the code or not. True by default.

This will use prettyplease, so the resulting formatted code will not be affected by the local rustfmt version or config.

pub fn emit_cargo_directives(self, emit_cargo_directives: bool) -> Abigen

Specify whether to print cargo build script directives if the source is a path. By default, this is true only when executing inside of a build script.

pub fn generate(self) -> Result<ContractBindings, Report>

Generates the contract bindings.

pub fn expand(self) -> Result<(ExpandedContract, Context), Report>

Expands the Abigen and returns the ExpandedContract that holds all tokens and the Context that holds the state used during expansion.

impl Abigen

pub fn source(&self) -> &Source

Returns a reference to the contract’s ABI source.

pub fn source_mut(&mut self) -> &mut Source

Returns a mutable reference to the contract’s ABI source.

pub fn name(&self) -> &Ident

Returns a reference to the contract’s name.

pub fn name_mut(&mut self) -> &mut Ident

Returns a mutable reference to the contract’s name.

pub fn method_aliases(&self) -> &HashMap<String, String, RandomState>

Returns a reference to the contract’s method aliases.

pub fn method_aliases_mut( &mut self ) -> &mut HashMap<String, String, RandomState>

Returns a mutable reference to the contract’s method aliases.

pub fn event_aliases(&self) -> &HashMap<String, String, RandomState>

Returns a reference to the contract’s event aliases.

pub fn error_aliases_mut(&mut self) -> &mut HashMap<String, String, RandomState>

Returns a mutable reference to the contract’s event aliases.

pub fn derives(&self) -> &Vec<Path, Global>

Returns a reference to the contract’s derives.

pub fn derives_mut(&mut self) -> &mut Vec<Path, Global>

Returns a mutable reference to the contract’s derives.