Module ethers_contract::multicall_contract::multicall_3

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

Aggregate3Call

Container type for all input parameters for the aggregate3 function with signature aggregate3((address,bool,bytes)[]) and selector 0x82ad56cb

Aggregate3Return

Container type for all return fields from the aggregate3 function with signature aggregate3((address,bool,bytes)[]) and selector 0x82ad56cb

Aggregate3ValueCall

Container type for all input parameters for the aggregate3Value function with signature aggregate3Value((address,bool,uint256,bytes)[]) and selector 0x174dea71

Aggregate3ValueReturn

Container type for all return fields from the aggregate3Value function with signature aggregate3Value((address,bool,uint256,bytes)[]) and selector 0x174dea71

AggregateCall

Container type for all input parameters for the aggregate function with signature aggregate((address,bytes)[]) and selector 0x252dba42

AggregateReturn

Container type for all return fields from the aggregate function with signature aggregate((address,bytes)[]) and selector 0x252dba42

BlockAndAggregateCall

Container type for all input parameters for the blockAndAggregate function with signature blockAndAggregate((address,bytes)[]) and selector 0xc3077fa9

BlockAndAggregateReturn

Container type for all return fields from the blockAndAggregate function with signature blockAndAggregate((address,bytes)[]) and selector 0xc3077fa9

Call

Call(address,bytes)

Call3

Call3(address,bool,bytes)

Call3Value

Call3Value(address,bool,uint256,bytes)

GetBasefeeCall

Container type for all input parameters for the getBasefee function with signature getBasefee() and selector 0x3e64a696

GetBasefeeReturn

Container type for all return fields from the getBasefee function with signature getBasefee() and selector 0x3e64a696

GetBlockHashCall

Container type for all input parameters for the getBlockHash function with signature getBlockHash(uint256) and selector 0xee82ac5e

GetBlockHashReturn

Container type for all return fields from the getBlockHash function with signature getBlockHash(uint256) and selector 0xee82ac5e

GetBlockNumberCall

Container type for all input parameters for the getBlockNumber function with signature getBlockNumber() and selector 0x42cbb15c

GetBlockNumberReturn

Container type for all return fields from the getBlockNumber function with signature getBlockNumber() and selector 0x42cbb15c

GetChainIdCall

Container type for all input parameters for the getChainId function with signature getChainId() and selector 0x3408e470

GetChainIdReturn

Container type for all return fields from the getChainId function with signature getChainId() and selector 0x3408e470

GetCurrentBlockCoinbaseCall

Container type for all input parameters for the getCurrentBlockCoinbase function with signature getCurrentBlockCoinbase() and selector 0xa8b0574e

GetCurrentBlockCoinbaseReturn

Container type for all return fields from the getCurrentBlockCoinbase function with signature getCurrentBlockCoinbase() and selector 0xa8b0574e

GetCurrentBlockDifficultyCall

Container type for all input parameters for the getCurrentBlockDifficulty function with signature getCurrentBlockDifficulty() and selector 0x72425d9d

GetCurrentBlockDifficultyReturn

Container type for all return fields from the getCurrentBlockDifficulty function with signature getCurrentBlockDifficulty() and selector 0x72425d9d

GetCurrentBlockGasLimitCall

Container type for all input parameters for the getCurrentBlockGasLimit function with signature getCurrentBlockGasLimit() and selector 0x86d516e8

GetCurrentBlockGasLimitReturn

Container type for all return fields from the getCurrentBlockGasLimit function with signature getCurrentBlockGasLimit() and selector 0x86d516e8

GetCurrentBlockTimestampCall

Container type for all input parameters for the getCurrentBlockTimestamp function with signature getCurrentBlockTimestamp() and selector 0x0f28c97d

GetCurrentBlockTimestampReturn

Container type for all return fields from the getCurrentBlockTimestamp function with signature getCurrentBlockTimestamp() and selector 0x0f28c97d

GetEthBalanceCall

Container type for all input parameters for the getEthBalance function with signature getEthBalance(address) and selector 0x4d2301cc

GetEthBalanceReturn

Container type for all return fields from the getEthBalance function with signature getEthBalance(address) and selector 0x4d2301cc

GetLastBlockHashCall

Container type for all input parameters for the getLastBlockHash function with signature getLastBlockHash() and selector 0x27e86d6e

GetLastBlockHashReturn

Container type for all return fields from the getLastBlockHash function with signature getLastBlockHash() and selector 0x27e86d6e

Multicall3

Result

Result(bool,bytes)

TryAggregateCall

Container type for all input parameters for the tryAggregate function with signature tryAggregate(bool,(address,bytes)[]) and selector 0xbce38bd7

TryAggregateReturn

Container type for all return fields from the tryAggregate function with signature tryAggregate(bool,(address,bytes)[]) and selector 0xbce38bd7

TryBlockAndAggregateCall

Container type for all input parameters for the tryBlockAndAggregate function with signature tryBlockAndAggregate(bool,(address,bytes)[]) and selector 0x399542e9

TryBlockAndAggregateReturn

Container type for all return fields from the tryBlockAndAggregate function with signature tryBlockAndAggregate(bool,(address,bytes)[]) and selector 0x399542e9

Enums

Multicall3Calls

Container type for all of the contract’s call

Statics

MULTICALL3_ABI

The parsed JSON ABI of the contract.