ethers_contract
Module stream

Contains the EventStream type which aids in streaming access to contract events

Structs

EventStream

Generic wrapper around Log streams, mapping their content to a specific deserialized log struct.

EventStreamMeta

Wrapper around a EventStream, that in addition to the deserialized Event type also yields the LogMeta.

SelectEvent

Stream for EventStream::select

Type Aliases

SelectEither

A stream of two items